+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.2.0 (2015-01-30)
+---------------------
+
+* modify ``loglevel_option`` (is_eager, callback with storing value on context)
+* add an abstract ``logger_option`` and a ``logfile_option``
+
 0.1.0 (2015-01-26)
 ---------------------